fix procces delete books

+++ app/Http/Controllers/BukuController.php
@@ -37,21 +37,21 @@ class BukuController extends Controller
     {
     	DB::table('buku')->where('id', $req->id)
                          ->update(
-                                    ['nama_buku' => $req->nama_buku,
-                                    'harga_buku' => $req->harga_buku,
-                                    'penulis' => $req->penulis,
-                                    'penerbit' => $req->penerbit,
-                                    'tahun' => $req->tahun]
+                                ['nama_buku' => $req->nama_buku,
+                                'harga_buku' => $req->harga_buku,
+                                'penulis' => $req->penulis,
+                                'penerbit' => $req->penerbit,
+                                'tahun' => $req->tahun]
                             );
     	return redirect('/buku')->with('success', 'Buku diedit');
     }
 
     public function destroy($id)
     {
-        // $status = Peminjaman::where('id_buku', $id)->where('status', '!=', 'dikembalikan')->count();
-        // if ($status > 0) {
-        //     return redirect('/buku')->with('danger', 'Buku masih dalam peminjaman');
-        // }
+        $status = Peminjaman::where('id_buku', $id)->where('status', '!=', 'dikembalikan')->count();
+        if ($status > 0) {
+            return redirect('/buku')->with('danger', 'Buku masih dalam peminjaman');
+        }
     	Buku::destroy($id);
     	return redirect('/buku')->with('success', 'Buku dihapus');
     }
